refactor(WinningNumbers): add static forTest() for Test
add static forTest() method for WinningNumbersTest Instance
